% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inp_options.R
\name{aermap_control_options}
\alias{aermap_control_options}
\title{AERMAP (CO) Control Options}
\usage{
aermap_control_options(
  TITLETWO = list(NULL, "Subtitle")[[1]],
  CHECKS = list(NULL, TRUE, FALSE)[[1]],
  TIFFDEBUGS = list(NULL, "path(s)/to/debug/file(s)")[[1]],
  ElevUnits = list(NULL, "FEET", "DECIFEET", "DECAFEET", "METERS", "DECIMETERS",
    "DECAMETERS")[[1]],
  TERRHGTS = list(NULL, "EXTRACT", "PROVIDED")[[1]],
  FLAGPOLE = list(NULL, 0)[[1]],
  DOMAINXY = c(xmin = NULL, xmax = NULL, ymin = NULL, ymax = NULL, zone_min = NULL,
    zone_max = NULL),
  DOMAINLL = c(xmin = NULL, xmax = NULL, ymin = NULL, ymax = NULL),
  NADGRIDS = list(NULL, "path/to/NADGrid")[[1]],
  DEBUGOPT = list(NULL, "HILL", "RECEPTOR", "SOURCE", "ALL")[[1]],
  ...,
  .expand_paths = TRUE
)
}
\arguments{
\item{TITLETWO}{TITLETWO control option (default: NULL - AERMAP assumes no subtitle).
If not NULL, must be a character string of the subtitle for the model run.}

\item{CHECKS}{CHECK control option (default: NULL - AERMAP does not check the file).
Only used if terrain_data_type is "DEM".
If not NULL, must be one of TRUE or FALSE - TRUE indicates that the preprocessor should check the input data files for errors.}

\item{TIFFDEBUGS}{TIFFDEBUG control option (default: NULL - AERMAP does not check the file).
Only used if terrain_data_type is "NED".
If not NULL, must be one of TRUE or FALSE - TRUE indicates that the preprocessor should check the input data files for errors.}

\item{ElevUnits}{ElevUnit control option (default: NULL - AERMAP does not check the file).
Only used if TIFFDEBUG is TRUE and terrain_data_type is "NED".
If not NULL, must be one of "METERS", "FEET", "DECIFEET", "DECAFEET", "DECIMETERS", "DECAMETERS" - indicates the elevation unit of the input data files.}

\item{TERRHGTS}{TERRHGTS control option (default: NULL - AERMAP assumes "EXTRACT").
If not NULL, must be one of "EXTRACT" or "PROVIDED".
If "EXTRACT", instructs the preprocessor to determine the terrain heights from the DEM data files provided by the user.
If "PROVIDED", forces the preprocessor to use the user-specified receptor elevations that are entered on the receptor pathway
(also applies to source elevations specified on the optional source pathway)}

\item{FLAGPOLE}{FLAGPOLE control option (default: NULL - AERMAP assumes 0).
Default numeric value (>= 0) for height of (flagpole) receptors above local ground level.}

\item{DOMAINXY}{DOMAINXY control option (default: NULL - AERMAP assumes full domain of input data).
If not NULL, must be a numeric vector of UTM coordinates (x, y, and zone) for the lower left and upper right corners of the domain
with named elements xmin, xmax, ymin, ymax, zone_min, and zone_max.}

\item{DOMAINLL}{DOMAINLL control option (default: NULL - AERMAP assumes full domain of input data).
If not NULL, must be a numeric vector of longitude and latitude coordinates for the lower left and upper right corners of the domain
with named elements xmin, xmax, ymin, and ymax.}

\item{NADGRIDS}{NADGRIDS control option (default: NULL - AERMAP assumes the same directory as the AERMAP .inp file).
If not NULL, must be a character string of the path to the folder containing NADCON grid shift files (.las and .los).}

\item{DEBUGOPT}{DEBUGOPT control option (default: NULL - AERMAP does not print debug information).
If not NULL, must be a character value equal to "ALL" (equivalent to \code{c("HILL", "RECEPTOR", "SOURCE")}), or a vector of 1 or more of "HILL", "RECEPTOR", or "SOURCE".
Separate debug files are created for each of these options to document the determination of the:
critical hill height scales,
receptor elevations and coordinate conversions,
and source elevations and coordinate conversions}

\item{...}{Additional named args (for future-proofing)}

\item{.expand_paths}{Should symbolic links etc in the paths be expanded? (default: TRUE)}
}
\value{
A list of set control options for AERMAP.
}
\description{
AERMAP (CO) Control Options
}
\references{
U.S. Environmental Protection Agency (EPA). (2018).
\emph{AERMAP User's Guide} (Version 18081), Table B-2.
https://gaftp.epa.gov/aqmg/SCRAM/models/related/aermap/aermap_userguide_v18081.pdf
}
