% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aermap_options.R
\name{aermap_output_options}
\alias{aermap_output_options}
\title{AERMAP output options}
\usage{
aermap_output_options(
  DEBUGHIL = list(NULL, "path/to/HILL/debug/file")[[1]],
  DEBUGREC = list(NULL, c("path/to/NADconv&DomainCheck/file",
    "path/to/NEDassignment/file", "path/to/RECelev/file"))[[1]],
  DEBUGSRC = list(NULL, c("path/to/NADconv&DomainCheck/file",
    "path/to/NEDassignment/file", "path/to/SRCelev/file"))[[1]],
  ...
)
}
\arguments{
\item{DEBUGHIL}{DEBUGHIL control option (default: NULL - AERMAP does not print HILL debug information).
If not NULL, must be a character string of the path to the HILL debug file.}

\item{DEBUGREC, DEBUGSRC}{DEBUGREC/DEBUGSRC control option (default: NULL - AERMAP does not print RECEPTOR/SOURCE debug information).
If not NULL, must be a character vector with the paths to the receptor/source NAD conversion and domain check file, the DEM-NED assignment file, and the receptor/source elevation file, in that order.}

\item{...}{Additional named args (for future-proofing)}
}
\value{
A list of set output options for AERMAP.
}
\description{
AERMAP output options
}
