% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aermet_options.R
\name{aermet_job_options}
\alias{aermet_job_options}
\title{AERMET Job Pathway Optional Settings}
\usage{
aermet_job_options(
  REPORT = NULL,
  CHK_SYNTAX = FALSE,
  NOPRINT = FALSE,
  DEBUG = NULL,
  ...
)
}
\arguments{
\item{REPORT}{Summary report file path (default: NULL)
If NULL, then the summary is written to \strong{"the output control device connected to logical unit 6. On a personal computer, this unit is normally the video monitor."}}

\item{CHK_SYNTAX}{Whether to check the syntax of the input file prior to processing (default: FALSE)}

\item{NOPRINT}{Whether to suppress the output of the model to the active session (default: FALSE)}

\item{DEBUG}{Debug file path (default: NULL)
If NULL, no debug file is created.
Note, the debug file can get large depending on the model inputs.}

\item{...}{Additional named args (for future-proofing)}
}
\description{
AERMET Job Pathway Optional Settings
}
