% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aermet_options.R
\name{aermet_metprep_options}
\alias{aermet_metprep_options}
\title{AERMET MetPrep Pathway Optional Settings}
\usage{
aermet_metprep_options(
  MODEL = list(NULL, "AERMOD")[[1]],
  THRESH_1MIN = list(NULL, 0.5)[[1]],
  LOCATION = list(NULL, list(site_id = "123456", lat = 50.123, lng = -175.123, tz_offset
    = 7))[[1]],
  XDATES = list(NULL, as.Date(c("2000-01-01", "2000-01-02")))[[1]],
  METHOD = NULL,
  UAWINDOW = list(NULL, list(start = -1, end = 1))[[1]],
  AERSURF = list(NULL, "path/to/file")[[1]],
  AERSURF2 = list(NULL, "path/to/file")[[1]],
  SSTDEPTH = list(NULL, list(SST = 0, DEPTH = 0))[[1]],
  ZI_GUST = list(NULL, 600)[[1]],
  ZI_MIN = list(NULL, 25)[[1]],
  MIN_MOL = list(NULL, 5)[[1]],
  DEF_VPTG = list(NULL, 0.01)[[1]],
  ...
)
}
\arguments{
\item{MODEL}{Model name (default: NULL - assumes "AERMOD").}

\item{THRESH_1MIN}{Threshold  wind speed for the 1-minute ASOS data (default: NULL - assumes 0.5 m/s).}

\item{LOCATION}{Location used for time conversion -
only needed when site-specific mixing heights are provided during Stage 1 and
upper air sounding data are omitted from the processing (default: NULL).
If not NULL, must be a list with elements "site_id", "lat", "lng", and optionally "tz_offset".}

\item{XDATES}{Date range (Date objects) to extract from the archive file (default: NULL)
If NULL, all data are extracted from the archive file if \code{EXTRACT} is not NULL.}

\item{METHOD}{define processing methods for the input data including data substitution,
special treatment of ASOS wind data, stable boundary layer treatment, and upper air sounding selection.
(default: NULL).
If not NULL, must be a list with one or more entries named
"WIND_DIR", "REFLEVEL", "ASOS_ADJ", "UASELECT", "STABLEBL", "CCVR", "TEMP", and/or "COARE",
each of which must be a vector with one or more named logical parameters (see details).}

\item{UAWINDOW}{Window for selecting upper air sounding data as a list with elements "start" and "end"
with the hour difference relative to the reference sounding times (default: NULL - equivelant to \code{list(start = -1, end = 1)}).}

\item{AERSURF, AERSURF2}{Primary/secondary surface characteristic file paths (default: NULL).}

\item{SSTDEPTH}{Sea surface temperature and measurement depth (default: NULL).
If not null, must be a list with elements "SST" and "DEPTH" specifying
the sea surface temperature (in degrees C) and depth of measurement (in metres).}

\item{ZI_GUST, ZI_MIN}{override the default mixing height to use for gust calculations and
the default minimum mixing height used in COARE. Units are metres. (defaults: NULL - equivelant to 600 and 25)}

\item{MIN_MOL}{minimum absolute value of Monin-Obukov length allowed. Units are metres. (default: NULL - equivelant to 5).}

\item{DEF_VPTG}{default value of potential temperature gradient output by COARE. Units are degrees C per metre. (default: NULL - equivelant to 0.01).}

\item{...}{Additional named args (for future-proofing)}
}
\description{
AERMET MetPrep Pathway Optional Settings
}
\details{
METHOD parameters: # TODO:
}
