% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aermet_options.R
\name{aermet_onsite_prog_options}
\alias{aermet_onsite_prog_options}
\title{AERMET Onsite or Prognostic Pathway Optional Settings}
\usage{
aermet_onsite_prog_options(
  XDATES = list(NULL, as.Date(c("2000-01-01", "2000-01-02")))[[1]],
  QAOUT = list(NULL, "path/to/file")[[1]],
  AUDIT = list(NULL, c("var1", "var2"), "ALL")[[1]],
  RANGE = list(NULL, data.frame(name = "var1", min = 1, symbol = "<=", max = 2,
    missing_indicator = "-9"))[[1]],
  NO_MISSING = list(NULL, c("var1", "var2"), "ALL")[[1]],
  OSHEIGHTS = list(NULL, c(1, 2, 3))[[1]],
  DELTA_TEMP = list(NULL, data.frame(lower = c(0, 10, 100), upper = c(10, 100,
    1000)))[[1]],
  OBS_HOUR = list(NULL, 6),
  ...
)
}
\arguments{
\item{XDATES}{Date range (Date objects) to extract from the archive file (default: NULL)
If NULL, all data are extracted from the archive file if \code{EXTRACT} is not NULL.}

\item{QAOUT}{Quality assurance output file path (default: NULL).
If NULL, no QA is performed and no output is written.
If not NULL, dry bulb temperature, wind speed, and wind direction (as well as any variables specified in \code{AUDIT}) are QA'd and summarized to the QAOUT file.}

\item{AUDIT}{Additional variables to QA or "ALL" for all variables (default: NULL).
If NULL, no additional QA is performed.}

\item{RANGE}{Upper/lower limits and missing value indicators of variables for QA (default: NULL).
If NULL, defaults defined in Table B-1 of the AERMET User's Guide are used (https://gaftp.epa.gov/Air/aqmg/SCRAM/models/met/aermet/aermet_userguide.pdf).}

\item{NO_MISSING}{Whether to suppress missing value indicators in QA output for specified variables (default: NULL).}

\item{OSHEIGHTS}{an altnerate way to provide measurement heights for multi-level profile data (default: NULL).
If not NULL, must be a vector of measurement heights (in meters) for each level of the profile.}

\item{DELTA_TEMP}{Defines the lower and upper measurement heights (m) for onsite temperature difference variable(s) (DT01–DT03),
which represent directly measured temperature gradients rather than differences between independent temperature sensors.
Only DT01 (index = 1) is used in later processing stages of AERMET (default: NULL).
If not NULL, must be a data.frame with columns "lower" and "upper" of measurement heights (in meters) for each level of the profile (to a maximum of 3 levels).}

\item{OBS_HOUR}{"OBS/HOUR" parameter - defines how many observations per hour for sub-hourly data (default: NULL).}

\item{...}{Additional named args (for future-proofing)}
}
\description{
AERMET Onsite or Prognostic Pathway Optional Settings
}
