% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aermet_options.R
\name{aermet_surface_options}
\alias{aermet_surface_options}
\title{AERMET Surface Pathway Optional Settings}
\usage{
aermet_surface_options(
  EXTRACT = list(NULL, "path/to/file")[[1]],
  XDATES = list(NULL, as.Date(c("2000-01-01", "2000-01-02")))[[1]],
  QAOUT = list(NULL, "path/to/file")[[1]],
  AUDIT = list(NULL, c("var1", "var2"), "ALL")[[1]],
  RANGE = list(NULL, data.frame(name = "var1", min = 1, symbol = "<=", max = 2,
    missing_indicator = "-9"))[[1]],
  NO_MISSING = list(NULL, c("var1", "var2"), "ALL")[[1]],
  ASOS1MIN = list(NULL, "path/to/file")[[1]],
  ...
)
}
\arguments{
\item{EXTRACT}{Where to write data retrieved from the archive file (default: NULL)
If NULL, no data is extracted from the archive file.}

\item{XDATES}{Date range (Date objects) to extract from the archive file (default: NULL)
If NULL, all data are extracted from the archive file if \code{EXTRACT} is not NULL.}

\item{QAOUT}{Quality assurance output file path (default: NULL).
If NULL, no QA is performed and no output is written.
If not NULL, dry bulb temperature, wind speed, and wind direction (as well as any variables specified in \code{AUDIT}) are QA'd and summarized to the QAOUT file.}

\item{AUDIT}{Additional variables to QA or "ALL" for all variables (default: NULL).
If NULL, no additional QA is performed.}

\item{RANGE}{Upper/lower limits and missing value indicators of variables for QA (default: NULL).
If NULL, defaults defined in Table B-1 of the AERMET User's Guide are used (https://gaftp.epa.gov/Air/aqmg/SCRAM/models/met/aermet/aermet_userguide.pdf).}

\item{NO_MISSING}{Whether to suppress missing value indicators in QA output for specified variables (default: NULL).}

\item{ASOS1MIN}{Path to the ASOS1MIN file if using wind speed and wind direction derived from 1-minute or 5-minute ASOS wind observations available from the NCEI (default: NULL).}

\item{...}{Additional named args (for future-proofing)}
}
\description{
AERMET Surface Pathway Optional Settings
}
