\name{AERrMOD-package}
\alias{AERrMOD-package}
\alias{AERrMOD}
\docType{package}
\title{
What the package does (short line)

}
\description{
More about what it does (maybe more than one line)

}
\details{
\tabular{ll}{
Package: \tab AERrMOD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-24\cr
License: \tab What license is it under?\cr
}

}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>

}
\references{

}

\keyword{ package }
\seealso{


}
\examples{

}
