% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_installed.R
\name{check_installed}
\alias{check_installed}
\title{Check if AERMOD is installed}
\usage{
check_installed(
  path = NULL,
  exe_name = ifelse(.Platform$OS.type == "windows", "aermod.exe", "aermod"),
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to AERMOD executable (optional). If provided and does not end in \code{exe_name}, it will be appended.}

\item{exe_name}{Name of the AERMOD executable (default: "aermod.exe" on Windows, "aermod" otherwise)}

\item{verbose}{Whether to print a success message if AERMOD is found (default: TRUE)}
}
\value{
A list containing the found path, the source of the path (either "user", "PATH", or "common_path"), and whether the path was found.
}
\description{
Check if AERMOD is installed
}
