% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create a new AERMOD R project}
\usage{
create_project(
  path,
  models = c("aermod", "aermap", "aermet"),
  remove_zips = TRUE,
  verbose = TRUE,
  prompt = rlang::is_interactive(),
  open = rlang::is_interactive(),
  ...
)
}
\arguments{
\item{path}{The path to the project directory to be created.}

\item{models}{A character vector of models to include in the project. The
models must be one of "aermod", "aermap", or "aermet".}

\item{remove_zips}{Whether to remove the downloaded zip file afterwards (default: TRUE)}

\item{verbose}{Whether to print a success message after downloading and unzipping (default: TRUE)}

\item{prompt}{Whether to prompt the user before downloading/unzipping (default: TRUE if R is interactive)}

\item{open}{Whether to open the project directory in the file explorer after
creation.}
}
\value{
The path to the project directory.
}
\description{
Create a new AERMOD R project
}
