% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_and_unzip}
\alias{get_and_unzip}
\title{Download and unzip a file}
\usage{
get_and_unzip(
  zip_url,
  local_dir,
  remove_zips = TRUE,
  verbose = TRUE,
  prompt = rlang::is_interactive()
)
}
\arguments{
\item{zip_url}{URL of the zip file to download}

\item{local_dir}{Local directory to download and unzip the file to}

\item{remove_zips}{Whether to remove the downloaded zip file afterwards (default: TRUE)}

\item{verbose}{Whether to print a success message after downloading and unzipping (default: TRUE)}

\item{prompt}{Whether to prompt the user before downloading/unzipping (default: TRUE if R is interactive)}
}
\value{
The path to the unzipped directory
}
\description{
Download and unzip a file
}
