% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_model.R
\name{install_model}
\alias{install_model}
\title{Download and unzip AERMOD, AERMAP, or AERMET model executables}
\usage{
install_model(
  path = ".",
  model = c("aermod", "aermap", "aermet")[1],
  is_x32 = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to the model executable (optional). If provided and does not end in \code{exe_name}, it will be appended.}

\item{model}{Name of the model (default: "aermod", must be one of "aermod", "aermap", or "aermet")}

\item{is_x32}{Whether to download the 32-bit version of the model (default: FALSE)}

\item{verbose}{Whether to print a success message if the model is already installed (default: TRUE)}
}
\value{
The path to the model executable. The model will be installed in \verb{\{path\}/\{toupper(model)\}/}.
}
\description{
Download and unzip AERMOD, AERMAP, or AERMET model executables
}
