% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aermet_inp.R
\name{make_aermet_inp}
\alias{make_aermet_inp}
\title{Create an AERMET input file}
\usage{
make_aermet_inp(
  inp_path = "aermet.inp",
  surface_station,
  instrument_heights,
  upperair_station,
  onsite_prog_station = NULL,
  site_characteristics = data.frame(frequency = list(NULL, c("ANNUAL", "SEASONAL",
    "MONTHLY"))[[1]], years = list(NULL, c("2020 2021 2022", "2023", "2024 2025"))[[1]],
    sector_start = NULL, sector_end = NULL, frequency_id = NULL, sector_id = NULL, albedo
    = NULL, bowen_ratio = NULL, surface_roughness = NULL),
  messages_path = "aermet_messages.txt",
  output_path = sub(basename(inp_path), pattern = "\\\\..*$", replacement = ".sfc"),
  profile_path = sub(basename(inp_path), pattern = "\\\\..*$", replacement = "_QA.pfl"),
  surface_path = sub(basename(inp_path), pattern = "\\\\..*$", replacement = ".iqa"),
  surface_data_format = c("EXTRACT", "CD144", "SCRAM", "SAMSON", "HUSWO", "ISHD")[1],
  upperair_path = sub(basename(inp_path), pattern = "\\\\..*$", replacement =
    "_upperair.txt"),
  upperair_data_format = c("EXTRACT", "IGRA", "FSL", "6201FB", "6201VB")[1],
  onsite_prog_path = sub(basename(inp_path), pattern = "\\\\..*$", replacement =
    "_onsite_prog.txt"),
  onsite_formats = NULL,
  windspeed_zero_threshold = 0,
  job_options = aermet_job_options(),
  surface_options = aermet_surface_options(),
  upperair_options = aermet_upperair_options(),
  onsite_prog_options = aermet_onsite_prog_options(),
  metprep_options = aermet_metprep_options(),
  is_asos = FALSE,
  use_onsite = TRUE,
  onsite_prog_over_land = TRUE,
  expand_paths = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{inp_path}{Path to input file (default: "aermet.inp").}

\item{surface_station, upperair_station, onsite_prog_station}{Station location details - must be a data.frame or list with names
"site_id", "lat", "lng", and optionally "tz_offset" and/or "elev".
\code{onsite_prog_station} is only required if using an onsite or prognostic station.}

\item{instrument_heights}{Heights of surface instruments (in meters) with names matching the instrument variable name (default: NULL).}

\item{site_characteristics}{Data.frame of site characteristics with columns
"frequency", "years", "sector_start", "sector_end", "frequency_id", "sector_id", "albedo", "bowen_ratio", and "surface_roughness".}

\item{messages_path}{Path to messages file (default: "aermet_messages.txt").}

\item{output_path, profile_path}{Paths to output surface and profile files passed to the OUTPUT and PROFILE options.}

\item{surface_path, upperair_path, onsite_prog_path}{Paths to archive files passed to each DATA option.}

\item{surface_data_format, upperair_data_format}{Formats of archive files passed to each DATA option.
For surface, must be one of "EXTRACT", "CD144", "SCRAM", "SAMSON", "HUSWO", or "ISHD".
For upperair, must be one of "EXTRACT", "IGRA", "FSL", "6201FB", "6201VB".}

\item{onsite_formats}{List of onsite data formats (default: NULL). See  Table B-3 and B-4 of the AERMET user guide.}

\item{windspeed_zero_threshold}{Threshold for the minimum wind speed required to detect air flow (default: 0 m/s).}

\item{job_options, surface_options, upperair_options, onsite_prog_options, metprep_options}{Additional optional settings passed to each control pathway in the input file.
See \code{\link[=aermet_job_options]{aermet_job_options()}}, \code{\link[=aermet_surface_options]{aermet_surface_options()}}, \code{\link[=aermet_upperair_options]{aermet_upperair_options()}}, \code{\link[=aermet_onsite_prog_options]{aermet_onsite_prog_options()}}, and \code{\link[=aermet_metprep_options]{aermet_metprep_options()}}.}

\item{is_asos}{Whether or not the surface station is an ASOS station (default: FALSE).}

\item{use_onsite}{Whether to use an onsite or prognostic pathway for the onsite_prog option (default: TRUE).}

\item{onsite_prog_over_land}{Whether or not the onsite/prog station is over land (instad of over water) (default: TRUE).}

\item{expand_paths}{Whether to expand relative paths to absolute paths (default: TRUE).}

\item{verbose}{Whether to print extra messages (default: TRUE).}
}
\description{
Create an AERMET input file
}
