% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aermap.R
\name{run_aermap}
\alias{run_aermap}
\title{Run AERMAP model}
\usage{
run_aermap(
  path = "AERMAP",
  exe_name = "aermap.exe",
  run_name = "aermap",
  inp_name = paste0(run_name, ".inp"),
  out_name = paste0(run_name, ".out"),
  log_name = paste0(run_name, ".log"),
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to directory with the model executable (default: ./AERMAP/).}

\item{exe_name}{Name of the model executable (default: "aermap.exe" on Windows, "aermap" otherwise)}

\item{run_name}{Name of the run (default: "aermap") - used for default names of the input and output files}

\item{inp_name}{Name of the input file (default: "{run_name}.inp")}

\item{out_name}{Name of the output file (default: "{run_name}.out")}

\item{log_name}{Name of the log file (default: "{run_name}.log")}

\item{verbose}{Whether to print a success message if AERMAP is found (default: TRUE)}
}
\value{
A character value of the model log (invisible), or an error if the model log is not found (i.e. the model failed to run)
}
\description{
Run AERMAP model
}
