% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aermet.R
\name{run_aermet}
\alias{run_aermet}
\title{Run AERMET model}
\usage{
run_aermet(
  path = "AERMET",
  exe_name = "aermet.exe",
  run_names = c("aermet1", "aermet2"),
  inp_names = paste0(run_names, ".inp"),
  log_names = paste0(run_names, ".log"),
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to directory with the model executable.}

\item{exe_name}{Name of the model executable (default: "\code{model}.exe")}

\item{run_names}{Names of the runs (default: c("aermet1", "aermet2")) - used for default names of the input and output files for each stage.}

\item{inp_names}{Names of the input files (default: run_names + ".inp")}

\item{log_names}{Names of the log files (default: run_names + ".log")}

\item{verbose}{Whether to print the output of the model (default: TRUE)}
}
\value{
A list of length 2 with character values of the model log for each stage (invisible), or an error if the model log is not found (i.e. the model failed to run)
}
\description{
A wrapper for \code{\link[=run_model]{run_model()}} to run both stages of the AERMET model.
}
