% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aermod.R
\name{run_aermod}
\alias{run_aermod}
\title{Run AERMOD model}
\usage{
run_aermod(
  path = "AERMOD",
  exe_name = "aermod.exe",
  run_name = "aermod",
  inp_name = paste0(run_name, ".inp"),
  out_name = paste0(run_name, ".out"),
  log_name = paste0(run_name, ".log"),
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to directory with the model executable.}

\item{exe_name}{Name of the model executable (default: "\code{model}.exe")}

\item{run_name}{Name of the run (default: \code{model}) - used for default names of the input and output files}

\item{inp_name}{Name of the input file (default: "\code{run_name}.inp")}

\item{out_name}{Name of the output file (default: "\code{run_name}.out")}

\item{log_name}{Name of the log file (default: "\code{run_name}.log")}

\item{verbose}{Whether to print the output of the model (default: TRUE)}
}
\value{
A character value of the model log (invisible), or an error if the model log is not found (i.e. the model failed to run)
}
\description{
A wrapper for \code{\link[=run_model]{run_model()}} to run the AERMOD model.
}
