% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run an AERMOD-related model}
\usage{
run_model(
  path,
  model = c("aermap", "aermet", "aermod")[1],
  exe_name = paste0(model, ".exe"),
  run_name = model,
  inp_name = paste0(run_name, ".inp"),
  out_name = paste0(run_name, ".out"),
  log_name = paste0(run_name, ".log"),
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to directory with the model executable.}

\item{model}{Name of model to run (default: "aermap", must be one of "aermod", "aermap", or "aermet")}

\item{exe_name}{Name of the model executable (default: "{model}.exe")}

\item{run_name}{Name of the run (default: {model}) - used for default names of the input and output files}

\item{inp_name}{Name of the input file (default: "{run_name}.inp")}

\item{out_name}{Name of the output file (default: "{run_name}.out")}

\item{log_name}{Name of the log file (default: "{run_name}.log")}

\item{verbose}{Whether to print the output of the model (default: TRUE)}
}
\value{
A character value of the model log (invisible), or an error if the model log is not found (i.e. the model failed to run)
}
\description{
Run an AERMOD-related model
}
