% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_case.R
\name{setup_test_case}
\alias{setup_test_case}
\title{Setup test case}
\usage{
setup_test_case(
  path,
  remove_zips = TRUE,
  verbose = TRUE,
  prompt = rlang::is_interactive(),
  open = rlang::is_interactive()
)
}
\arguments{
\item{path}{The path to the project directory to be created.}

\item{remove_zips}{Whether to remove the downloaded zip file afterwards (default: TRUE)}

\item{verbose}{Whether to print a success message after downloading and unzipping (default: TRUE)}

\item{prompt}{Whether to prompt the user before downloading/unzipping (default: TRUE if R is interactive)}

\item{open}{Whether to open the project directory in the file explorer after
creation.}
}
\value{
The path to the test case directory, with the zip file removed if
remove_zips is TRUE.
}
\description{
This function sets up a test case for the aermodr package
by downloading and unzipping the SampleRun.zip file from the
US EPA Support Center for Regulatory Atmospheric Modeling (SCRAM) website.
}
