% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_installed.R
\name{check_installed}
\alias{check_installed}
\title{Check if an AERMOD-related model is installed}
\usage{
check_installed(
  path = NULL,
  model = c("aermod", "aermap", "aermet")[1],
  exe_name = paste0(model, ifelse(.Platform$OS.type == "windows", ".exe", "")),
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to the model executable (optional). If provided and does not end in \code{exe_name}, it will be appended.}

\item{model}{Name of the model (default: "aermod", must be one of "aermod", "aermap", or "aermet")}

\item{exe_name}{Name of the model executable (default: "\code{model}.exe" on Windows, "\code{model}" otherwise)}

\item{verbose}{Whether to print a success message if the model is found (default: TRUE)}
}
\value{
A list containing the found path, the source of the path (either "user", "PATH", or "common_path"), and whether the path was found.
}
\description{
Check if an AERMOD-related model is installed
}
