% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_model.R
\name{install_model}
\alias{install_model}
\title{Download and unzip AERMOD, AERMAP, or AERMET model executables}
\usage{
install_model(
  path = ".",
  model = c("aermod", "aermap", "aermet")[1],
  is_x32 = FALSE,
  remove_zips = TRUE,
  verbose = TRUE,
  prompt = FALSE
)
}
\arguments{
\item{path}{Path to the model executable (optional). If provided and does not end in \code{exe_name}, it will be appended.}

\item{model}{Name of the model (default: "aermod", must be one of "aermod", "aermap", or "aermet")}

\item{is_x32}{Whether to download the 32-bit version of the model (default: FALSE)}

\item{remove_zips}{Whether to remove the downloaded zip file afterwards (default: TRUE)}

\item{verbose}{Whether to print a success message if the model is found (default: TRUE)}

\item{prompt}{Whether to prompt the user before downloading/unzipping (default: TRUE if R is interactive)}
}
\value{
The path to the model executable. The model will be installed in \verb{\{path\}/\{toupper(model)\}/}.
}
\description{
Note: for AERMAP, the .los and .las files normally within the AERMAP directory are moved into "NAD_gridshifts/" after unxzipping.
Ensure that the input file for AERMAP sets NADGRIDS in the control options to "NAD_gridshifts/" (the default if using \code{\link[=make_aermap_inp]{make_aermap_inp()}}).
}
